requirements.txt
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
peewee==3.10.0
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.19
python-dotenv==0.10.3
six==1.12.0
Werkzeug==0.15.5
